Move GenderSelector to /wca
